requirements.txt
python-telegram-bot==21.1
pandas
openpyxl